requirements.txt
# BagunçArt Requirements
kivy>=2.2.0
kivymd>=1.2.0
mysql-connector-python>=8.0.33
bcrypt>=4.0.1
python-dotenv>=1.0.0
requests>=2.31.0
python-dateutil>=2.8.2
pillow>=9.0.0

# Development
buildozer
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
